Bootstrapped measures of precision

superb also comes with a few built-in measures of precisions that uses bootstrap. More can be added based on users needs. All ``bootstrapSE.fct()`` functions produces an interval width; all ``bootstrapCI.fct()`` produces the lower and upper limits of an interval. These estimates are based on 5,000 sub-samples by default. Change this default with``options("superb.bootstrapIter" = number )``. See (Efron and Tibshirani 1994) for a comprehensive introduction. All "superbPlot-compatible" precision measures must have these parameters:

bootstrapSE.mean(x)

bootstrapCI.mean(x, gamma)

bootstrapSE.median(x)

bootstrapCI.median(x, gamma)

bootstrapSE.hmean(x)

bootstrapCI.hmean(x, gamma)

bootstrapSE.gmean(x)

bootstrapCI.gmean(x, gamma)

bootstrapSE.var(x)

bootstrapCI.var(x, gamma)

bootstrapSE.sd(x)

bootstrapCI.sd(x, gamma)

Arguments

x
a vector of numbers, the sample data (mandatory);

gamma
a confidence level for CI (default 0.95).

Value

a measure of precision (SE) or an interval of precision (CI).

References

Efron B, Tibshirani RJ (1994). An introduction to the bootstrap. CRC press.

Examples

# the confidence interval of the mean for default 95% and 90% confidence level
bootstrapCI.mean( c(1,2,3) )

#>  2.5% 97.5% 
#>     1     3 

bootstrapCI.mean( c(1,2,3), gamma = 0.90)

#>       5%      95% 
#> 1.333333 2.666667 

# Standard errors for standard deviation or variance
bootstrapSE.sd( c(1,2,3) )

#> [1] 0.3543689

bootstrapSE.var( c(1,2,3) )

#> [1] 0.4702706