Additional summary statistics

superb adds a few summary statistics that can be used to characterize a dataset. All comes with SE.fct() and CI.fct(). See (Harding et al. 2014; Harding et al. 2015) for more. superbPlot-compatible summary statistics functions must have one parameter:

hmean(x)

gmean(x)

MAD(x)

fisherskew(x)

pearsonskew(x)

fisherkurtosis(x)

Arguments

x
a vector of numbers, the sample data (mandatory);

Value

a summary statistic describing the sample.

References

Harding B, Tremblay C, Cousineau D (2014). “Standard errors: A review and evaluation of standard error estimators using Monte Carlo simulations.” The Quantitative Methods for Psychology, 10(2), 107--123.

Harding B, Tremblay C, Cousineau D (2015). “The standard error of the Pearson skew.” The Quantitative Methods for Psychology, 11(1), 32--36.

Examples

# the confidence interval of the mean for default 95% and 90% confidence level
gmean( c(1,2,3) )  # the geometric mean; also available in psych::geometric.mean   

#> [1] 1.817121

hmean( c(1,2,3) )  # the harmonic mean;  also available in psych::harmonic.mean   

#> [1] 1.636364

MAD( c(1,2,3) )    # the median absolute deviation to the median (not the same as mad)

#> [1] 1

fisherskew( c(1,2,3) )     # the Fisher skew corrected for sample size

#> [1] 0

fisherkurtosis( c(1,2,3) ) # the Fisher kurtosis corrected for sample size

#> [1] NaN

pearsonskew( c(1,2,3) )    # the Pearson skew

#> [1] 0