superbData to obtain summary statistics with correct error bar intervals.

plotsuberb plots standard error or confidence interval for various descriptive statistics under various designs, sampling schemes, population size and purposes, according to the suberb framework. See (Cousineau 2017) for more.

superbData(
  data,
  BSFactor = NULL,
  WSFactor = NULL,
  factorOrder,
  variables,
  statistic = "mean",
  errorbar = "CI",
  gamma = 0.95,
  adjustments = list(purpose = "single", popSize = Inf, decorrelation = "none",
    samplingDesign = "SRS"),
  preprocessfct = NULL,
  postprocessfct = NULL,
  clusterColumn = ""
)

Arguments

data
Dataframe in wide format

BSFactor
The name of the columns containing the between-subject factor(s)

WSFactor
The name of the within-subject factor(s)

factorOrder
Order of factors as shown in the graph (x axis, groups, horizontal panels, vertical panels)

variables
The dependent variable(s)

statistic
The summary statistic function to use

errorbar
The function that computes the error bar. Should be "CI" or "SE" or any function name. Defaults to "SE"

gamma
The coverage factor; necessary when errorbar == "CI". Default is 0.95.

adjustments
List of adjustments as described below: popsize: Size of the population under study. Defaults to Inf purpose: The purpose of the comparisons. Defaults to "single". Can be "single", "difference" or "tryon". decorrelation: Decorrelation method for repeated measure designs. Chooses among the methods ("CM", "LM", "CA" or "none"). Defaults to "none". samplingDesign: Sampling method to obtain the sample. implemented sampling is "SRS" (Simple Randomize Sampling) and "CRS" (Cluster-Randomized Sampling). Default is adjustments = list(purpose = "single", popSize = Inf, decorrelation = "none", samplingDesign = "SRS")

preprocessfct
is a transform (or vector of) to be performed first on data matrix of each group

postprocessfct
is a transform (or vector of)

clusterColumn
used in conjunction with samplingDesign = "CRS", indicates which column contains the cluster membership

Value

a list with (1) the summary statistics in summaryStatistics (2) the raw data in long format, with standardized levels in rawData.

References

Cousineau D (2017). “Varieties of confidence intervals.” Advances in Cognitive Psychology, 13, 140--155. doi: 10.5709/acp-0214-z , https://doi.org/10.5709/acp-0214-z.

Examples

# basic example using a built-in dataframe as data; 
# by default, the mean is computed and the error bar are 95% confidence intervals
superbData(ToothGrowth, BSFactor = c("dose", "supp"), 
  variables = "len") 

#> $summaryStatistics
#>   dose supp  mean - 95 % CI width + 95 % CI width
#> 1  0.5   OJ 13.23       -3.190283        3.190283
#> 2  0.5   VC  7.98       -1.964824        1.964824
#> 3    1   OJ 22.70       -2.797727        2.797727
#> 4    1   VC 16.77       -1.799343        1.799343
#> 5    2   OJ 26.06       -1.899314        1.899314
#> 6    2   VC 26.14       -3.432090        3.432090
#> 
#> $rawData
#>    supp dose id   DV
#> 1    VC  0.5  1  4.2
#> 2    VC  0.5  2 11.5
#> 3    VC  0.5  3  7.3
#> 4    VC  0.5  4  5.8
#> 5    VC  0.5  5  6.4
#> 6    VC  0.5  6 10.0
#> 7    VC  0.5  7 11.2
#> 8    VC  0.5  8 11.2
#> 9    VC  0.5  9  5.2
#> 10   VC  0.5 10  7.0
#> 11   VC    1 11 16.5
#> 12   VC    1 12 16.5
#> 13   VC    1 13 15.2
#> 14   VC    1 14 17.3
#> 15   VC    1 15 22.5
#> 16   VC    1 16 17.3
#> 17   VC    1 17 13.6
#> 18   VC    1 18 14.5
#> 19   VC    1 19 18.8
#> 20   VC    1 20 15.5
#> 21   VC    2 21 23.6
#> 22   VC    2 22 18.5
#> 23   VC    2 23 33.9
#> 24   VC    2 24 25.5
#> 25   VC    2 25 26.4
#> 26   VC    2 26 32.5
#> 27   VC    2 27 26.7
#> 28   VC    2 28 21.5
#> 29   VC    2 29 23.3
#> 30   VC    2 30 29.5
#> 31   OJ  0.5 31 15.2
#> 32   OJ  0.5 32 21.5
#> 33   OJ  0.5 33 17.6
#> 34   OJ  0.5 34  9.7
#> 35   OJ  0.5 35 14.5
#> 36   OJ  0.5 36 10.0
#> 37   OJ  0.5 37  8.2
#> 38   OJ  0.5 38  9.4
#> 39   OJ  0.5 39 16.5
#> 40   OJ  0.5 40  9.7
#> 41   OJ    1 41 19.7
#> 42   OJ    1 42 23.3
#> 43   OJ    1 43 23.6
#> 44   OJ    1 44 26.4
#> 45   OJ    1 45 20.0
#> 46   OJ    1 46 25.2
#> 47   OJ    1 47 25.8
#> 48   OJ    1 48 21.2
#> 49   OJ    1 49 14.5
#> 50   OJ    1 50 27.3
#> 51   OJ    2 51 25.5
#> 52   OJ    2 52 26.4
#> 53   OJ    2 53 22.4
#> 54   OJ    2 54 24.5
#> 55   OJ    2 55 24.8
#> 56   OJ    2 56 30.9
#> 57   OJ    2 57 26.4
#> 58   OJ    2 58 27.3
#> 59   OJ    2 59 29.4
#> 60   OJ    2 60 23.0
#> 

# example changing the summary statistics to the median and
# the error bar to 90% confidence intervals
superbData(ToothGrowth, BSFactor = c("dose", "supp"), 
  variables = "len", statistic = "median", errorbar = "CI", gamma = .90) 

#> $summaryStatistics
#>   dose supp median - 90 % CI width + 90 % CI width
#> 1  0.5   OJ  12.25       -3.240079        3.240079
#> 2  0.5   VC   7.15       -1.995492        1.995492
#> 3    1   OJ  23.45       -2.841396        2.841396
#> 4    1   VC  16.50       -1.827429        1.827429
#> 5    2   OJ  25.95       -1.928960        1.928960
#> 6    2   VC  25.95       -3.485660        3.485660
#> 
#> $rawData
#>    supp dose id   DV
#> 1    VC  0.5  1  4.2
#> 2    VC  0.5  2 11.5
#> 3    VC  0.5  3  7.3
#> 4    VC  0.5  4  5.8
#> 5    VC  0.5  5  6.4
#> 6    VC  0.5  6 10.0
#> 7    VC  0.5  7 11.2
#> 8    VC  0.5  8 11.2
#> 9    VC  0.5  9  5.2
#> 10   VC  0.5 10  7.0
#> 11   VC    1 11 16.5
#> 12   VC    1 12 16.5
#> 13   VC    1 13 15.2
#> 14   VC    1 14 17.3
#> 15   VC    1 15 22.5
#> 16   VC    1 16 17.3
#> 17   VC    1 17 13.6
#> 18   VC    1 18 14.5
#> 19   VC    1 19 18.8
#> 20   VC    1 20 15.5
#> 21   VC    2 21 23.6
#> 22   VC    2 22 18.5
#> 23   VC    2 23 33.9
#> 24   VC    2 24 25.5
#> 25   VC    2 25 26.4
#> 26   VC    2 26 32.5
#> 27   VC    2 27 26.7
#> 28   VC    2 28 21.5
#> 29   VC    2 29 23.3
#> 30   VC    2 30 29.5
#> 31   OJ  0.5 31 15.2
#> 32   OJ  0.5 32 21.5
#> 33   OJ  0.5 33 17.6
#> 34   OJ  0.5 34  9.7
#> 35   OJ  0.5 35 14.5
#> 36   OJ  0.5 36 10.0
#> 37   OJ  0.5 37  8.2
#> 38   OJ  0.5 38  9.4
#> 39   OJ  0.5 39 16.5
#> 40   OJ  0.5 40  9.7
#> 41   OJ    1 41 19.7
#> 42   OJ    1 42 23.3
#> 43   OJ    1 43 23.6
#> 44   OJ    1 44 26.4
#> 45   OJ    1 45 20.0
#> 46   OJ    1 46 25.2
#> 47   OJ    1 47 25.8
#> 48   OJ    1 48 21.2
#> 49   OJ    1 49 14.5
#> 50   OJ    1 50 27.3
#> 51   OJ    2 51 25.5
#> 52   OJ    2 52 26.4
#> 53   OJ    2 53 22.4
#> 54   OJ    2 54 24.5
#> 55   OJ    2 55 24.8
#> 56   OJ    2 56 30.9
#> 57   OJ    2 57 26.4
#> 58   OJ    2 58 27.3
#> 59   OJ    2 59 29.4
#> 60   OJ    2 60 23.0
#> 

# example introducing adjustments for pairwise comparisons 
# and assuming that the whole population is limited to 200 persons
superbData(ToothGrowth, BSFactor = c("dose", "supp"), 
  variables = "len",  
  adjustments = list( purpose = "difference", popSize = 200) )

#> $summaryStatistics
#>   dose supp  mean - 95 % CI width + 95 % CI width
#> 1  0.5   OJ 13.23       -4.397502        4.397502
#> 2  0.5   VC  7.98       -2.708323        2.708323
#> 3    1   OJ 22.70       -3.856401        3.856401
#> 4    1   VC 16.77       -2.480224        2.480224
#> 5    2   OJ 26.06       -2.618024        2.618024
#> 6    2   VC 26.14       -4.730810        4.730810
#> 
#> $rawData
#>    supp dose id   DV
#> 1    VC  0.5  1  4.2
#> 2    VC  0.5  2 11.5
#> 3    VC  0.5  3  7.3
#> 4    VC  0.5  4  5.8
#> 5    VC  0.5  5  6.4
#> 6    VC  0.5  6 10.0
#> 7    VC  0.5  7 11.2
#> 8    VC  0.5  8 11.2
#> 9    VC  0.5  9  5.2
#> 10   VC  0.5 10  7.0
#> 11   VC    1 11 16.5
#> 12   VC    1 12 16.5
#> 13   VC    1 13 15.2
#> 14   VC    1 14 17.3
#> 15   VC    1 15 22.5
#> 16   VC    1 16 17.3
#> 17   VC    1 17 13.6
#> 18   VC    1 18 14.5
#> 19   VC    1 19 18.8
#> 20   VC    1 20 15.5
#> 21   VC    2 21 23.6
#> 22   VC    2 22 18.5
#> 23   VC    2 23 33.9
#> 24   VC    2 24 25.5
#> 25   VC    2 25 26.4
#> 26   VC    2 26 32.5
#> 27   VC    2 27 26.7
#> 28   VC    2 28 21.5
#> 29   VC    2 29 23.3
#> 30   VC    2 30 29.5
#> 31   OJ  0.5 31 15.2
#> 32   OJ  0.5 32 21.5
#> 33   OJ  0.5 33 17.6
#> 34   OJ  0.5 34  9.7
#> 35   OJ  0.5 35 14.5
#> 36   OJ  0.5 36 10.0
#> 37   OJ  0.5 37  8.2
#> 38   OJ  0.5 38  9.4
#> 39   OJ  0.5 39 16.5
#> 40   OJ  0.5 40  9.7
#> 41   OJ    1 41 19.7
#> 42   OJ    1 42 23.3
#> 43   OJ    1 43 23.6
#> 44   OJ    1 44 26.4
#> 45   OJ    1 45 20.0
#> 46   OJ    1 46 25.2
#> 47   OJ    1 47 25.8
#> 48   OJ    1 48 21.2
#> 49   OJ    1 49 14.5
#> 50   OJ    1 50 27.3
#> 51   OJ    2 51 25.5
#> 52   OJ    2 52 26.4
#> 53   OJ    2 53 22.4
#> 54   OJ    2 54 24.5
#> 55   OJ    2 55 24.8
#> 56   OJ    2 56 30.9
#> 57   OJ    2 57 26.4
#> 58   OJ    2 58 27.3
#> 59   OJ    2 59 29.4
#> 60   OJ    2 60 23.0
#> 

# This example is based on repeated measures
library(lsr)
options(superb.feedback = 'none') # shut down 'warnings' and 'design' interpretation messages

# define shorter column names...
names(Orange) <- c("Tree","age","circ")
# turn the data into a wide format
Orange.wide <- longToWide(Orange, circ ~ age)
superbData( Orange.wide, WSFactor = "age(7)",
  variables = c("circ_118","circ_484","circ_664","circ_1004","circ_1231","circ_1372","circ_1582"),
  adjustments = list(purpose = "difference", decorrelation = "none")
)

#> $summaryStatistics
#>   age  mean - 95 % CI width + 95 % CI width
#> 1   1  31.0       -2.483328        2.483328
#> 2   2  57.8      -14.341085       14.341085
#> 3   3  93.2      -30.272166       30.272166
#> 4   4 134.2      -45.543856       45.543856
#> 5   5 145.6      -51.324448       51.324448
#> 6   6 173.4      -57.661892       57.661892
#> 7   7 175.8      -58.442686       58.442686
#> 
#> $rawData
#>    Tree age  DV
#> 1     1   1  30
#> 2     2   1  33
#> 3     3   1  30
#> 4     4   1  32
#> 5     5   1  30
#> 6     1   2  58
#> 7     2   2  69
#> 8     3   2  51
#> 9     4   2  62
#> 10    5   2  49
#> 11    1   3  87
#> 12    2   3 111
#> 13    3   3  75
#> 14    4   3 112
#> 15    5   3  81
#> 16    1   4 115
#> 17    2   4 156
#> 18    3   4 108
#> 19    4   4 167
#> 20    5   4 125
#> 21    1   5 120
#> 22    2   5 172
#> 23    3   5 115
#> 24    4   5 179
#> 25    5   5 142
#> 26    1   6 142
#> 27    2   6 203
#> 28    3   6 139
#> 29    4   6 209
#> 30    5   6 174
#> 31    1   7 145
#> 32    2   7 203
#> 33    3   7 140
#> 34    4   7 214
#> 35    5   7 177
#> 

superbData( Orange.wide, WSFactor = "age(7)",
  variables = c("circ_118","circ_484","circ_664","circ_1004","circ_1231","circ_1372","circ_1582"),
  adjustments = list(purpose = "difference", decorrelation = "CM")
)

#> $summaryStatistics
#>   age  mean - 95 % CI width + 95 % CI width
#> 1   1  31.0       -36.59726        36.59726
#> 2   2  57.8       -29.27023        29.27023
#> 3   3  93.2       -13.24987        13.24987
#> 4   4 134.2       -11.25778        11.25778
#> 5   5 145.6       -16.76345        16.76345
#> 6   6 173.4       -24.46650        24.46650
#> 7   7 175.8       -25.42244        25.42244
#> 
#> $rawData
#>    Tree age  DV
#> 1     1   1  30
#> 2     2   1  33
#> 3     3   1  30
#> 4     4   1  32
#> 5     5   1  30
#> 6     1   2  58
#> 7     2   2  69
#> 8     3   2  51
#> 9     4   2  62
#> 10    5   2  49
#> 11    1   3  87
#> 12    2   3 111
#> 13    3   3  75
#> 14    4   3 112
#> 15    5   3  81
#> 16    1   4 115
#> 17    2   4 156
#> 18    3   4 108
#> 19    4   4 167
#> 20    5   4 125
#> 21    1   5 120
#> 22    2   5 172
#> 23    3   5 115
#> 24    4   5 179
#> 25    5   5 142
#> 26    1   6 142
#> 27    2   6 203
#> 28    3   6 139
#> 29    4   6 209
#> 30    5   6 174
#> 31    1   7 145
#> 32    2   7 203
#> 33    3   7 140
#> 34    4   7 214
#> 35    5   7 177
#>