User Interface to get summary plot of any statistics with adjusted error bars.

The function suberbShiny() provides a simple user interface to plot standard error or confidence interval for various descriptive statistics under various designs, population size and purposes, according to the suberb framework. See (Cousineau et al. 2021) for more. Also see this video from (Walker 2021) for a demo using the shinyapps.io installation accessible at dcousin3.shinyapps.io/superbshiny Limitations: it is not possible to use custom-made statistics with the graphical user interface, nor is it possible to request an adjustment for cluster- randomized sampling. These options are available with superbPlot().

superbShiny()

Value

A plot that can be cut-and-paste.

References

Cousineau D, Goulet M, Harding B (2021). “Summary plots with adjusted error bars: The superb framework with an implementation in R.” Advances in Methods and Practices in Psychological Science, in press, 1--46. doi: 10.1177/25152459211035109 .

Walker JAL (2021). Summary plots with adjusted error bars (superb). https://www.youtube.com/watch?v=rw_6ll5nVus.

Examples

# Launch the user interface: 
# \donttest{
if (interactive())
   superbShiny() 
# }