two-step transform for subject centering and bias correction

twoStepTransform is a transformation that can be applied to a matrix of data. The resulting matrix is both subject-centered and bias corrected, a technique called the CM technique (Baguley 2012; Cousineau 2005; Morey 2008)

twoStepTransform(dta, variables)

Arguments

dta

a data.frame containing the data in wide format;

variables

a vector of column names on which the transformation will be applied. the remaining columns will be left unchanged

Value

a data.frame of the same form as dta with the variables transformed.

This function is useful when passed to the argument preprocessfct of superbPlot()

where it performs a modification of the data matrix.

References

Baguley T (2012). “Calculating and graphing within-subject confidence intervals for ANOVA.” Behavior Research Methods, 44, 158 -- 175. doi: 10.3758/s13428-011-0123-7 .

Cousineau D (2005). “Confidence intervals in within-subject designs: A simpler solution to Loftus and Masson's method.” Tutorials in Quantitative Methods for Psychology, 1, 42 -- 45. doi: 10.20982/tqmp.01.1.p042 .

Morey RD (2008). “Confidence Intervals from Normalized Data: A correction to Cousineau (2005).” Tutorials in Quantitative Methods for Psychology, 4, 61 -- 64. doi: 10.20982/tqmp.04.2.p061 .