Bootstrapped measures of precision

superb also comes with a few built-in measures of precisions that uses bootstrap. More can be added based on users needs. All bootstrapSE.fct() functions produces an interval width; all bootstrapPI.fct() produces the lower and upper limits of an interval. These estimates are based on 5,000 sub-samples by default. Change this default withoptions("superb.bootstrapIter" = number ). See Efron and Tibshirani (1994) for a comprehensive introduction. The bootstrap estimates are called PI which stands for Precision intervals. This is to denote that they estimate the sampling distribution, not the predictive distribution on which all confidence intervals are based (Rousselet et al. 2019; Poitevineau and Lecoutre 2010; Lecoutre 1999) .

bootstrapSE.mean(x)

bootstrapPI.mean(x, gamma)

bootstrapSE.median(x)

bootstrapPI.median(x, gamma)

bootstrapSE.hmean(x)

bootstrapPI.hmean(x, gamma)

bootstrapSE.gmean(x)

bootstrapPI.gmean(x, gamma)

bootstrapSE.var(x)

bootstrapPI.var(x, gamma)

bootstrapSE.sd(x)

bootstrapPI.sd(x, gamma)

Arguments

x

a vector of numbers, the sample data (mandatory);

gamma

a confidence level for PI (default 0.95).

Value

a measure of precision (SE) or an interval of precision (PI).

References

Efron B, Tibshirani RJ (1994). An introduction to the bootstrap. CRC press.

Lecoutre B (1999). “Two useful distributions for Bayesian predictive procedures under normal models.” Journal of Statistical Planning and Inference, 79, 93 -- 105. doi:10.1016/S0378-3758(98)00231-6 .

Poitevineau J, Lecoutre B (2010). “Implementing Bayesian predictive procedures: The K-prime and K-square distributions.” Computational Statistics and Data Analysis, 54, 724 -- 731. doi:10.1016/j.csda.2008.11.004 .

Rousselet GA, Pernet CR, Wilcox RR (2019). “A practical introduction to the bootstrap: A versatile method to make inferences by using data-driven simulations.” psyArXiv. doi:10.31234/osf.io/h8ft7 .

Examples

# the confidence interval of the mean for default 95% and 90% confidence level
bootstrapPI.mean( c(1,2,3) )
#>  2.5% 97.5% 
#>     1     3 
bootstrapPI.mean( c(1,2,3), gamma = 0.90)
#>       5%      95% 
#> 1.333333 2.666667 

# Standard errors for standard deviation or variance
bootstrapSE.sd( c(1,2,3) )
#> [1] 0.3543689
bootstrapSE.var( c(1,2,3) )
#> [1] 0.4702706